Give middleware attachments their own module
An attached policy value has its own typing, collection validation, and
structured diagnostics, distinct from declaring a middleware spec. Keep that
cohesive API in MiddlewareAttachment instead of growing MiddlewareSpec into
both declaration and attachment management.

Move attachment-owned tests with the implementation and update consumers to
use the new public namespace and subpath. Spec compatibility rules and group
traversal retain their existing owners, and the attachment module depends on
MiddlewareSpec only through types so the split adds no runtime cycle.
